Refactor dataControllers
