+++ _common/gameModel/libs/ClientScript/edition/UIfacade/activableFacade.ts
@@ -4,7 +4,7 @@
 
 import { IActivableDescriptor, Uid } from '../../game/common/interfaces';
 import { entries } from '../../tools/helper';
-import { FlatActivable } from '../controllers/dataController';
+import { FlatActivable } from '../controllers/dataControllerBase';
 import { getFlatObjects } from '../UIfacade/genericConfigFacade';
 
 type ActivableTypeNames = FlatActivable['activableType'];
