+++ _common/gameModel/libs/ClientScript/edition/UIfacade/genericConfigFacade.ts
@@ -11,7 +11,7 @@ import {
   FlatTypeBySuperType,
   FlatTypes,
   SuperTypeNames,
-} from '../controllers/dataController';
+} from '../controllers/dataControllerBase';
 import { getCurrentPage } from './mainMenuStateFacade';
 
 /**
