+++ _common/gameModel/libs/ClientScript/edition/UIfacade/locationConfigFacade.ts
@@ -6,11 +6,12 @@ import { MapConfig } from '../../gameMap/utils/mapConfig';
 import { patchX } from '../../tools/helper';
 import { getLocationTranslation } from '../../UIfacade/locationFacade';
 import { getMapEntityController } from '../controllers/controllerInstances';
-import { MapEntityCreationOptions, MapEntityFlatType } from '../controllers/dataController';
+import { MapEntityFlatType } from '../controllers/dataControllerBase';
 import { FlatMapEntity } from '../typeDefinitions/mapEntityDefinition';
 import { FlatMapObject } from '../typeDefinitions/mapObjectDefinition';
 import { addNew, getItems, ModalState } from '../UIfacade/genericConfigFacade';
 import { GenericScenaristInterfaceState } from './genericConfigFacade';
+import { MapEntityCreationOptions } from '../controllers/mapEntityController';
 
 export type SupportedDrawType = Exclude<DrawType, 'Circle'>;
